FNOformer/requirements.txt
torch
torchaudio
neuraloperator
prodigyopt
soundfile
tqdm